one step left to add backward compatability for properties file
